src/main/java/Nikolai023/stationList/client/datatypes/Station.java
package Nikolai023.stationList.client.datatypes;

import com.google.gwt.user.client.rpc.IsSerializable;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

public class Station implements IsSerializable {
    private String name;
    private String address;
    private String phoneNumber;
    private List<String> services = new ArrayList<String>();

    public String getName() {
        return name;
    }

    public String getAddress() {
        return address;
    }

    public String getPhoneNumber() {
        return phoneNumber;
    }

    public List<String> getServices() {
        return services;
    }

    public void setName(String name) {
        this.name = name;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public void setPhoneNumber(String phoneNumber) {
        this.phoneNumber = phoneNumber;
    }

    public void setServices(List<String> services) {
        this.services = services;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;

        Station station = (Station) o;

        if (name != null ? !name.equals(station.name) : station.name != null) return false;
        if (address != null ? !address.equals(station.address) : station.address != null) return false;
        if (phoneNumber != null ? !phoneNumber.equals(station.phoneNumber) : station.phoneNumber != null) return false;
        return services != null ? Arrays.deepEquals(services.toArray(), station.services.toArray()) : station.services == null;
    }

    @Override
    public int hashCode() {
        int result = name != null ? name.hashCode() : 0;
        result = 31 * result + (address != null ? address.hashCode() : 0);
        result = 31 * result + (phoneNumber != null ? phoneNumber.hashCode() : 0);
        result = 31 * result + (services != null ? services.hashCode() : 0);
        return result;
    }
}
